Move tests for unsupported operations to a new test class
